Update AsyncConverter.nuspec
This changes the dependency on the `Wave` package to be exactly 7.0, and not any higher. Without this change, it is dependent on Wave >= 7.0, which means that when the package is uploaded to the plugin gallery, it is also visible to Wave 8.0. This means it's visible in the Extension Manager for 2017.1 EAP builds, but it's not compatible. To fix this:

* Change the dependency to Wave to be `"[7.0]"`
* Upload a new version with this fixed dependency and mark the previous version as "unlisted" in resharpen-plugins.jetbrains.com. Marking it unlisted will prevent the gallery from showing it to ReSharper 2017.1 EAP builds